requirements.txt
aioredis==2.0.0
aiounittest==1.4.1
alembic==1.7.5
anyio==3.5.0
asgiref==3.4.1
astroid==2.9.3
async-timeout==4.0.1
asyncpg==0.25.0
asynctest==0.13.0
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
debugpy==1.5.1
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.89.1
fastapi-cache2==0.1.8
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httptools==0.3.0
httpx==0.22.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
pendulum==2.1.2
platformdirs==2.4.1
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pylint==2.12.2
pylint-plugin-utils==0.7
pyparsing==3.0.6
pytest==6.2.5
pytest-asyncio==0.17.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytzdata==2020.1
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.46
starlette==0.22.0
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
watchgod==0.7
websockets==10.1
Werkzeug==2.0.2
wrapt==1.13.3
